requirements.txt
numpy
Pillow
PyYAML
scikit_image
torch
torchvision
